Remove bdkmanager, ldkmanager
